backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Reinaldo Silva <josereinaldo.pessoal@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
celery = "^5.4.0"
python-decouple = "^3.8"
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.31"
alembic = {extras = ["tz"], version = "^1.13.2"}
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
netmiko = {git = "https://github.com/ktbyers/netmiko.git", rev = "develop"}
ntc-templates = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
